refactor: refactor

+++ Assets/Scripts/Services/IBalanceService.cs
@@ -4,7 +4,7 @@ namespace Services
 {
      public interface IBalanceService : IService
      {
-          event Action<int> BalanceChanged;
+          event Action<float> BalanceChanged;
           void Pay(int amount);
           void Receive(int amount);
           bool HasEnoughMoney(int amount);
